aira-vector-search/requirements_vector.txt
fastapi
uvicorn
redis
openai
numpy
python-dotenv